작업하다 일하러감

+++ apps/src/main/kotlin/io/dustin/apps/board/domain/qna/question/repository/custom/CustomQuestionRepository.kt
@@ -4,5 +4,5 @@ import io.dustin.apps.board.domain.qna.question.model.dto.QuestionDto
 
 interface CustomQuestionRepository {
     fun getQuestion(userId: Long, questionId: Long): QuestionDto?
-    fun getQuestionList(userId: Long, nextId: Long?, size: Int): List<QuestionDto>
+    fun getQuestionList(userId: Long, nextId: Int?, size: Int): List<QuestionDto>
 }
