+++ apps/src/main/kotlin/io/dustin/apps/common/model/CountByPagingInfo.kt
@@ -1,3 +1,4 @@
 package io.dustin.apps.common.model
 
-data class CountByPagingInfo<T>(val result: List<T>, val isLast: Boolean, val nextId: Long)
+data class CountByPagingInfo<T>(
+    val result: List<T>, val isLast: Boolean, val nextId: Int)
